+++ app/src/main/AndroidManifest.xml
@@ -8,12 +8,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.VinMartAuth">
-        <activity android:name=".login.ResetPasswordActivity" />
-        <activity android:name=".login.VerificationActivity" />
-        <activity android:name=".login.SignupActivity" />
-        <activity android:name=".login.LoginActivity" />
-        <activity android:name=".MainActivity">
+        android:theme="@style/Theme.VinMartAuth"
+        android:fullBackupContent="@xml/backup_descriptor">
+        <activity android:name=".view.ResetPasswordActivity" />
+        <activity android:name=".view.VerificationActivity" />
+        <activity android:name=".view.SignupActivity" />
+        <activity android:name=".view.LoginActivity" />
+        <activity android:name=".view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
