+++ app/src/main/java/com/fpt/vinmartauth/adapter/ProfileMenuItemAdapter.java
@@ -1,4 +1,4 @@
-package com.fpt.vinmartauth;
+package com.fpt.vinmartauth.adapter;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,6 +8,9 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.fpt.vinmartauth.R;
+import com.fpt.vinmartauth.entity.ProfileMenuItem;
 
 import java.util.ArrayList;
