+++ app/src/main/java/com/fpt/vinmartauth/entity/ProfileMenuItem.java
@@ -1,4 +1,4 @@
-package com.fpt.vinmartauth;
+package com.fpt.vinmartauth.entity;
 
 public class ProfileMenuItem {
     int iconID;
