+++ app/src/main/java/com/fpt/vinmartauth/view/MainActivity.java
@@ -1,4 +1,4 @@
-package com.fpt.vinmartauth;
+package com.fpt.vinmartauth.view;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -11,6 +11,10 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
+import com.fpt.vinmartauth.R;
+import com.fpt.vinmartauth.view.fragment.CartFragment;
+import com.fpt.vinmartauth.view.fragment.MainFragment;
+import com.fpt.vinmartauth.view.fragment.ProfileFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 public class MainActivity extends AppCompatActivity {
