+++ app/src/main/java/com/fpt/vinmartauth/view/ResetPasswordActivity.java
@@ -1,4 +1,4 @@
-package com.fpt.vinmartauth.login;
+package com.fpt.vinmartauth.view;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,20 +9,20 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.fpt.vinmartauth.R;
 
-public class SignupActivity extends AppCompatActivity {
+public class ResetPasswordActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_signup);
+        setContentView(R.layout.activity_reset_password);
 
         ImageView btnBack = findViewById(R.id.btnBack);
         btnBack.setOnClickListener(view -> finish());
 
-        Button btnSignUp = findViewById(R.id.btnSignUp);
-        btnSignUp.setOnClickListener(view -> {
-            Intent intent = new Intent(this, VerificationActivity.class);
-            startActivity(intent);
+        Button btnResetPassword = findViewById(R.id.btnResetPassword);
+        btnResetPassword.setOnClickListener(view -> {
+            Intent i = new Intent(this, LoginActivity.class);
+            startActivity(i);
         });
     }
 }
