+++ app/src/main/java/com/fpt/vinmartauth/view/VerificationActivity.java
@@ -1,4 +1,4 @@
-package com.fpt.vinmartauth.login;
+package com.fpt.vinmartauth.view;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,20 +9,20 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.fpt.vinmartauth.R;
 
-public class ResetPasswordActivity extends AppCompatActivity {
+public class VerificationActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_reset_password);
+        setContentView(R.layout.activity_verification);
 
         ImageView btnBack = findViewById(R.id.btnBack);
         btnBack.setOnClickListener(view -> finish());
 
-        Button btnResetPassword = findViewById(R.id.btnResetPassword);
-        btnResetPassword.setOnClickListener(view -> {
-            Intent i = new Intent(this, LoginActivity.class);
-            startActivity(i);
+        Button btnVerify = findViewById(R.id.btnVerify);
+        btnVerify.setOnClickListener(view -> {
+            Intent intent = new Intent(this, MainActivity.class);
+            startActivity(intent);
         });
     }
 }
