+++ app/src/main/java/com/fpt/vinmartauth/view/fragment/CartFragment.java
@@ -1,4 +1,4 @@
-package com.fpt.vinmartauth;
+package com.fpt.vinmartauth.view.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,6 +6,8 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
+
+import com.fpt.vinmartauth.R;
 
 /**
  * A simple {@link Fragment} subclass.
