+++ app/src/main/res/layout/activity_reset_password.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/quite_white"
-    tools:context=".login.ResetPasswordActivity">
+    tools:context=".view.ResetPasswordActivity">
 
     <ImageView
         android:id="@+id/btnBack"
