+++ app/src/main/res/layout/activity_verification.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/quite_white"
-    tools:context=".login.VerificationActivity">
+    tools:context=".view.VerificationActivity">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/cardView"
